change: improve appium runner structure
