+++ e2e-tests/scripts/inject_secrets.sh
