+++ e2e-tests/scripts/resolve_ios_simulator.sh
