+++ e2e-tests/scripts/resolve_rn_cli_version.sh
