+++ e2e-tests/scripts/start_appium.sh
