+++ e2e-tests/template/MainActivity.kt.template
